Backend/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- SQL (PostgreSQL) ---
sqlmodel==0.0.22
psycopg[binary]==3.2.3
alembic==1.13.2

# --- NoSQL (MongoDB) ---
beanie==1.26.0
motor==3.5.1

# --- Config & utilities ---
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.2.0

# --- Security / auth ---
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# --- Domain utilities ---
qrcode[pil]==7.4.2

# --- Dev / testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
ruff==0.6.9
